backend-services/cancel_appointment/requirements.txt
functions-framework==3.*
psycopg2-binary
firebase-admin
google-cloud-pubsub